+++ src/helpers/Base.js
@@ -1,5 +1,4 @@
 import {
-  DATA_UI_PREFIX,
   EVENT_BREAKPOINT,
   ACTION_TOGGLE,
   ACTION_SHOW,
@@ -81,7 +80,7 @@ class Base {
       opts = mergeDeep(
         Default,
         getDataValue(_data, dataName, elem),
-        datasetValue,
+        isDataObject && datasetValue,
         opts,
       );
     }
@@ -128,6 +127,15 @@ class Base {
       {
         onUpdate: (breakpoint, opts) => {
           this.emit(EVENT_BREAKPOINT, breakpoint, this.breakpoint);
+          if (breakpoint[0].data) {
+            const dataValue = getDataValue(
+              this.constructor._data,
+              breakpoint[0].data,
+            );
+            if (dataValue) {
+              breakpoint[0] = { ...breakpoint[0], ...dataValue };
+            }
+          }
           this.opts = breakpoint[1]
             ? mergeDeep(opts, breakpoint[0])
             : { ...opts };